// src/components/Content/HomePage/InteractiveDiagram/data/flow-types.tsx
export type TooltipState = {
  visible: boolean;
  x: number;
  y: number;
  content: string;
};